Lightning Data: Refactor files (#19424)
(cherry picked from commit ac9d63f4eb3c6055c7fd67d2430e6d171e9a59aa)
