+++ src/lightning/data/__init__.py
@@ -1,7 +1,7 @@
+from lightning.data.processing.functions import map, optimize, walk
 from lightning.data.streaming.combined import CombinedStreamingDataset
 from lightning.data.streaming.dataloader import StreamingDataLoader
 from lightning.data.streaming.dataset import StreamingDataset
-from lightning.data.streaming.functions import map, optimize, walk
 
 __all__ = [
     "LightningDataset",
