+++ src/lightning/data/streaming/__init__.py
@@ -13,18 +13,14 @@
 
 from lightning.data.streaming.cache import Cache
 from lightning.data.streaming.combined import CombinedStreamingDataset
-from lightning.data.streaming.data_processor import DataChunkRecipe, DataProcessor, DataTransformRecipe
 from lightning.data.streaming.dataloader import StreamingDataLoader
 from lightning.data.streaming.dataset import StreamingDataset
 from lightning.data.streaming.item_loader import TokensLoader
 
 __all__ = [
     "Cache",
-    "DataProcessor",
     "StreamingDataset",
     "CombinedStreamingDataset",
     "StreamingDataLoader",
-    "DataTransformRecipe",
-    "DataChunkRecipe",
     "TokensLoader",
 ]
